Complete rework of phpgedview version 4.2.2 to use BitDBAdodb in place of PDO

+++ downloadgedcom.php
@@ -22,7 +22,7 @@
  *
  * @package PhpGedView
  * @subpackage Admin
- * @version $Id: downloadgedcom.php,v 1.11 2008/07/07 18:01:12 lsces Exp $
+ * @version $Id: downloadgedcom.php,v 1.12 2009/09/15 20:06:00 lsces Exp $
  */
 
 /**
@@ -38,29 +38,48 @@ $gGedcom = new BitGEDCOM();
 // leave manual config until we can move it to bitweaver table 
 require_once("config.php");
 require_once("includes/functions_export.php");
+// Validate user parameters
+if (!isset($_SESSION['exportConvPath'])) $_SESSION['exportConvPath'] = $MEDIA_DIRECTORY;
+if (!isset($_SESSION['exportConvSlashes'])) $_SESSION['exportConvSlashes'] = 'forward';
+
+$ged				= safe_GET('ged',				get_all_gedcoms());
+$action				= safe_GET('action',			'download');
+$remove				= safe_GET('remove',			'yes', 'no');
+$convert			= safe_GET('convert',			'yes', 'no');
+$zip				= safe_GET('zip',				'yes', 'no');
+$conv_path			= safe_GET('conv_path',			PGV_REGEX_NOSCRIPT,				$_SESSION['exportConvPath']);
+$conv_slashes		= safe_GET('conv_slashes',		array('forward', 'backward'),	$_SESSION['exportConvSlashes']);
+$privatize_export	= safe_GET('privatize_export',	array('none', 'visitor', 'user', 'gedadmin', 'admin'));
+$filetype			= safe_GET('filetype',			array('gedcom', 'gramps'));
+
+$conv_path = stripLRMRLM($conv_path);
+$_SESSION['exportConvPath'] = $conv_path;		// remember this for the next Download
+$_SESSION['exportConvSlashes'] = $conv_slashes;
+
 
 if ((!userGedcomAdmin(getUserName())) || (empty ($ged))) {
 	header("Location: editgedcoms.php");
 	exit;
 }
-if (!isset ($action))
-	$action = "";
-if (!isset ($remove))
-	$remove = "no";
-if (!isset ($convert))
-	$convert = "no";
-if (!isset ($zip))
-	$zip = "no";
-if (!isset ($privatize_export))
-	$privatize_export = "no";
+
+if ($action == 'download') {
+	$conv_path = rtrim(str_replace('\\', '/', trim($conv_path)), '/').'/';	// make sure we have a trailing slash here
+	if ($conv_path=='/') $conv_path = '';
+
+	$exportOptions = array();
+	$exportOptions['privatize'] = $privatize_export;
+	$exportOptions['toANSI'] = $convert;
+	$exportOptions['noCustomTags'] = $remove;
+	$exportOptions['path'] = $conv_path;
+	$exportOptions['slashes'] = $conv_slashes;
+}
 
 if ($action == "download" && $zip == "yes") {
 	require "includes/pclzip.lib.php";
 
 	$temppath = $INDEX_DIRECTORY . "tmp/";
 	$fileName = $ged;
-	if($filetype =="gramps")
-		$fileName = $ged.".gramps";
+	if ($filetype =="gramps") $fileName = $ged.".gramps";
 	$zipname = "dl" . date("YmdHis") . $fileName . ".zip";
 	$zipfile = $INDEX_DIRECTORY . $zipname;
 	$gedname = $temppath . $fileName;
@@ -77,23 +96,21 @@ if ($action == "download" && $zip == "yes") {
 	$gedout = fopen(filename_decode($gedname), "w");
 	switch ($filetype) {
 	case 'gedcom':
-		print_gedcom($privatize_export, $privatize_export_level, $convert, $remove, $gedout);
+		export_gedcom($GEDCOM, $gedout, $exportOptions);
 		break;
 	case 'gramps':
-		print_gramps($privatize_export, $privatize_export_level, $convert, $remove, $gedout);
+		export_gramps($GEDCOM, $gedout, $exportOptions);
 		break;
 	}
 	fclose($gedout);
-	$comment = "Created by PhpGedView " . $VERSION . " " . $VERSION_RELEASE . " on " . date("r") . ".";
+	$comment = "Created by ".PGV_PHPGEDVIEW." ".PGV_VERSION_TEXT." on " . date("r") . ".";
 	$archive = new PclZip(filename_decode($zipfile));
 	$v_list = $archive->create(filename_decode($gedname), PCLZIP_OPT_COMMENT, $comment, PCLZIP_OPT_REMOVE_PATH, filename_decode($temppath));
-	if ($v_list == 0)
-		print "Error : " . $archive->errorInfo(true);
+	if ($v_list == 0) print "Error : " . $archive->errorInfo(true);
 	else {
 		unlink(filename_decode($gedname));
-		if ($removeTempDir)
-			rmdir(filename_decode($temppath));
-		header("Location: downloadbackup.php?fname=" . rawurlencode($zipname));
+		if ($removeTempDir) rmdir(filename_decode($temppath));
+		header("Location: ".encode_url("downloadbackup.php?fname={$zipname}", false));
 		exit;
 	}
 	exit;
@@ -102,70 +119,73 @@ if ($action == "download" && $zip == "yes") {
 if ($action == "download") {
 	header("Content-Type: text/plain; charset=$CHARACTER_SET");
 	// We could open "php://compress.zlib" to create a .gz file or "php://compress.bzip2" to create a .bz2 file
-	$fp=fopen('php://output', 'w');
+	$gedout = fopen('php://output', 'w');
 	switch ($filetype) {
 	case 'gedcom':
-		header("Content-Disposition: attachment; filename={$ged}");
-		print_gedcom($privatize_export, $privatize_export_level, $convert, $remove, $fp);
+		header('Content-Disposition: attachment; filename="'.$ged.'"');
+		export_gedcom($GEDCOM, $gedout, $exportOptions);
 		break;
 	case 'gramps':
-		header("Content-Disposition: attachment; filename={$ged}.gramps");
-		print_gramps($privatize_export, $privatize_export_level, $convert, $remove, $fp);
+		header('Content-Disposition: attachment; filename="'.$ged.'.gramps"');
+		export_gramps($GEDCOM, $gedout, $exportOptions);
 		break;
 	}
-	fclose($fp);
+	fclose($gedout);
 	exit;
 }
 
 print_header($pgv_lang["download_gedcom"]);
 
 ?>
-	<div class="center">
-	<h2><?php print $pgv_lang["download_gedcom"]; ?></h2>
-	<br />
-	<form name="convertform" method="post">
-		<input type="hidden" name="action" value="download" />
-		<input type="hidden" name="ged" value="<?php print $ged; ?>" />
-		<table class="list_table" border="0" align="center" valign="top">
-		<tr><td colspan="2" class="facts_label03" style="text-align:left;">
-		<?php print $pgv_lang["options"]; ?>
+<div class="center"><h2><?php print $pgv_lang["download_gedcom"]; ?></h2></div>
+<br />
+<form name="convertform" method="get">
+	<input type="hidden" name="action" value="download" />
+	<input type="hidden" name="ged" value="<?php print $ged; ?>" />
+	<table class="list_table width50" border="0" valign="top">
+	<tr><td colspan="2" class="facts_label03"><?php print $pgv_lang["options"]; ?></td></tr>
+	<tr><td class="descriptionbox width50 wrap"><?php print_help_link("file_type_help", "qm"); print $pgv_lang["choose_file_type"] ?></td>
+		<td class="optionbox">
+		<?php if ($TEXT_DIRECTION=='ltr') { ?>
+			<input type="radio" name="filetype" checked="checked" value="gedcom" />&nbsp;&nbsp;GEDCOM<br/><input type="radio" name="filetype" value="gramps" />&nbsp;&nbsp;Gramps XML
+		<?php } else { ?>
+			GEDCOM&nbsp;&nbsp;<?php print getLRM();?><input type="radio" name="filetype" checked="checked" value="gedcom" /><?php print getLRM();?><br />Gramps XML&nbsp;&nbsp;<?php print getLRM();?><input type="radio" name="filetype" value="gramps" /><?php print getLRM();?>
+		<?php } ?>
 		</td></tr>
-		<td class="descriptionbox wrap" align="left"><?php print $pgv_lang["choose_file_type"] ?></td>
-		<td class="optionbox" align="left"><input type="radio" name="filetype" checked="checked"  value="gedcom" />GEDCOM 
-		<?php print_help_link("def_gedcom_help", "qm"); ?>
-		<br/>
-		<input type="radio" name="filetype" value="gramps" />Gramps XML 
-		<?php print_help_link("def_gramps_help", "qm"); ?>
+	<tr><td class="descriptionbox width50 wrap"><?php print_help_link("download_zipped_help", "qm"); print $pgv_lang["zip_files"]; ?></td>
+		<td class="list_value"><input type="checkbox" name="zip" value="yes" checked="checked" /></td></tr>
+	<tr><td class="descriptionbox width50 wrap"><?php print_help_link("apply_privacy_help", "qm"); print $pgv_lang["apply_privacy"]; ?></td>
+		<td class="list_value">
+		<?php if (PGV_USER_IS_ADMIN) { ?>
+			<input type="radio" name="privatize_export" value="none" checked="checked" />&nbsp;&nbsp;<?php print $pgv_lang["none"]; ?><br />
+			<input type="radio" name="privatize_export" value="visitor" />&nbsp;&nbsp;<?php print $pgv_lang["visitor"]; ?><br />
+		<?php } else { ?>
+			<input type="radio" name="privatize_export" value="none" DISABLED />&nbsp;&nbsp;<?php print $pgv_lang["none"]; ?><br />
+			<input type="radio" name="privatize_export" value="visitor" checked="checked" />&nbsp;&nbsp;<?php print $pgv_lang["visitor"]; ?><br />
+		<?php } ?>
+		<input type="radio" name="privatize_export" value="user" />&nbsp;&nbsp;<?php print $pgv_lang["user"]; ?><br />
+		<input type="radio" name="privatize_export" value="gedadmin" />&nbsp;&nbsp;<?php print $pgv_lang["gedadmin"]; ?><br />
+		<input type="radio" name="privatize_export" value="admin"<?php if (!PGV_USER_IS_ADMIN) print " DISABLED"; ?> />&nbsp;&nbsp;<?php print $pgv_lang["siteadmin"]; ?>
 		</td></tr>
-		<tr><td class="list_label" style="padding: 5px; text-align:<?php if ($TEXT_DIRECTION == "ltr") print "left"; else print "right";?>; "><?php print $pgv_lang["utf8_to_ansi"]; ?></td>
-			<td class="list_value" style="padding: 5px; text-align:<?php if ($TEXT_DIRECTION == "ltr") print "left"; else print "right";?>; "><input type="checkbox" name="convert" value="yes" /><?php print_help_link("utf8_ansi_help", "qm"); ?></td></tr>
-		<tr><td class="list_label" style="padding: 5px; text-align:<?php if ($TEXT_DIRECTION == "ltr") print "left"; else print "right";?>; "><?php print $pgv_lang["remove_custom_tags"]; ?></td>
-			<td class="list_value" style="padding: 5px; text-align:<?php if ($TEXT_DIRECTION == "ltr") print "left"; else print "right";?>; "><input type="checkbox" name="remove" value="yes" checked="checked" /><?php print_help_link("remove_tags_help", "qm"); ?></td></tr>
-		<tr><td class="list_label" style="padding: 5px; text-align:<?php if ($TEXT_DIRECTION == "ltr") print "left"; else print "right";?>; "><?php print $pgv_lang["zip_files"]; ?></td>
-			<td class="list_value" style="padding: 5px; text-align:<?php if ($TEXT_DIRECTION == "ltr") print "left"; else print "right";?>; "><input type="checkbox" name="zip" value="yes" checked="checked" /><?php print_help_link("download_zipped_help", "qm"); ?></td></tr>
-		<tr><td class="list_label" valign="baseline" style="padding: 5px; text-align:<?php if ($TEXT_DIRECTION == "ltr") print "left"; else print "right";?>; "><?php print $pgv_lang["apply_privacy"]; ?>
-			<div id="privtext" style="display: none"></div>
-			</td>
-			<td class="list_value" style="padding: 5px; text-align:<?php if ($TEXT_DIRECTION == "ltr") print "left"; else print "right";?>; ">
-			<input type="checkbox" name="privatize_export" value="yes" onclick="expand_layer('privtext'); expand_layer('privradio');" />
-			<?php print_help_link("apply_privacy_help", "qm"); ?>
-			<div id="privradio" style="display: none"><br /><?php print $pgv_lang["choose_priv"]; ?><br />
-			<input type="radio" name="privatize_export_level" value="visitor" checked="checked" />
-			<?php print $pgv_lang["visitor"]; ?><br />
-			<input type="radio" name="privatize_export_level" value="user" /><?php print $pgv_lang["user"]; ?><br />
-			<input type="radio" name="privatize_export_level" value="gedadmin" /><?php print $pgv_lang["gedadmin"]; ?><br />
-			<input type="radio" name="privatize_export_level" value="admin" /><?php print $pgv_lang["siteadmin"]; ?><br />
-		</div></td>
-		</tr>
-		<tr><td class="facts_label03" colspan="2" style="padding: 5px; ">
-		<input type="submit" value="<?php print $pgv_lang["download_now"]; ?>" />
-		<input type="button" value="<?php print $pgv_lang["back"];?>" onclick="window.location='editgedcoms.php';"/></td></tr>
-		</table><br />
+	<tr><td class="descriptionbox width50 wrap"><?php print_help_link("utf8_ansi_help", "qm"); print $pgv_lang["utf8_to_ansi"]; ?></td>
+		<td class="list_value"><input type="checkbox" name="convert" value="yes" /></td></tr>
+	<tr><td class="descriptionbox width50 wrap"><?php print_help_link("remove_tags_help", "qm"); print $pgv_lang["remove_custom_tags"]; ?></td>
+		<td class="list_value"><input type="checkbox" name="remove" value="yes" checked="checked" /></td></tr>
+	<tr><td class="descriptionbox width50 wrap"><?php print_help_link("convertPath_help", "qm"); print $pgv_lang["convertPath"];?></td>
+		<td class="list_value"><input type="text" name="conv_path" size="30" value="<?php echo getLRM(), $conv_path, getLRM();?>" /></td></tr>
+	<tr><td class="descriptionbox width50 wrap"><?php print_help_link("convertSlashes_help", "qm"); print $pgv_lang["convertSlashes"];?></td>
+		<td class="list_value">
+		<input type="radio" name="conv_slashes" value="forward" <?php if ($conv_slashes=='forward') print "checked=\"checked\" "; ?>/>&nbsp;&nbsp;<?php print $pgv_lang["forwardSlashes"];?><br />
+		<input type="radio" name="conv_slashes" value="backward" <?php if ($conv_slashes=='backward') print "checked=\"checked\" "; ?>/>&nbsp;&nbsp;<?php print $pgv_lang["backSlashes"];?>
+		</td></tr>
+	<tr><td class="facts_label03" colspan="2">
+	<input type="submit" value="<?php print $pgv_lang["download_now"]; ?>" />
+	<input type="button" value="<?php print $pgv_lang["back"];?>" onclick="window.location='editgedcoms.php';"/></td></tr>
+	</table><br />
 	<br /><br />
-	</form>
-	<?php
+</form>
+<?php
 
 print $pgv_lang["download_note"] . "<br /><br /><br />\n";
-print "</div>";
 print_footer();
 ?>
